只使用lru
